Refactor (remove nn and include in bijections module to avoid circular import)
