+++ flowjax/bijections/masked_autoregressive.py
@@ -1,5 +1,7 @@
 """Masked autoregressive network and bijection."""
 
+from __future__ import annotations
+
 from collections.abc import Callable
 from functools import partial
 
@@ -8,11 +10,13 @@ import jax
 import jax.nn as jnn
 import jax.numpy as jnp
 from jax import Array
+from jax.typing import ArrayLike
 
 from flowjax.bijections.bijection import AbstractBijection
 from flowjax.bijections.jax_transforms import Vmap
-from flowjax.nn import autoregressive_mlp
-from flowjax.utils import get_ravelled_bijection_constructor
+from flowjax.masks import rank_based_mask
+from flowjax.utils import get_ravelled_pytree_constructor
+from flowjax.wrappers import Where
 
 
 class MaskedAutoregressive(AbstractBijection):
@@ -39,7 +43,7 @@ class MaskedAutoregressive(AbstractBijection):
     shape: tuple[int, ...]
     cond_shape: tuple[int, ...] | None
     transformer_constructor: Callable
-    autoregressive_mlp: eqx.nn.MLP
+    masked_autoregressive_mlp: eqx.nn.MLP
 
     def __init__(
         self,
@@ -57,9 +61,7 @@ class MaskedAutoregressive(AbstractBijection):
                 "Only unconditional transformers with shape () are supported.",
             )
 
-        constructor, transformer_init_params = get_ravelled_bijection_constructor(
-            transformer,
-        )
+        constructor, num_params = get_ravelled_pytree_constructor(transformer)
 
         if cond_dim is None:
             self.cond_shape = None
@@ -72,9 +74,9 @@ class MaskedAutoregressive(AbstractBijection):
             )
 
         hidden_ranks = jnp.arange(nn_width) % dim
-        out_ranks = jnp.repeat(jnp.arange(dim), transformer_init_params.size)
+        out_ranks = jnp.repeat(jnp.arange(dim), num_params)
 
-        amlp = autoregressive_mlp(
+        self.masked_autoregressive_mlp = masked_autoregressive_mlp(
             in_ranks,
             hidden_ranks,
             out_ranks,
@@ -83,26 +85,19 @@ class MaskedAutoregressive(AbstractBijection):
             key=key,
         )
 
-        # Initialise bias terms to match the provided transformer parameters
-        self.autoregressive_mlp = eqx.tree_at(
-            where=lambda t: t.layers[-1].bias,
-            pytree=amlp,
-            replace=jnp.tile(transformer_init_params, dim),
-        )
-
         self.transformer_constructor = constructor
         self.shape = (dim,)
         self.cond_shape = None if cond_dim is None else (cond_dim,)
 
     def transform(self, x, condition=None):
         nn_input = x if condition is None else jnp.hstack((x, condition))
-        transformer_params = self.autoregressive_mlp(nn_input)
+        transformer_params = self.masked_autoregressive_mlp(nn_input)
         transformer = self._flat_params_to_transformer(transformer_params)
         return transformer.transform(x)
 
     def transform_and_log_det(self, x, condition=None):
         nn_input = x if condition is None else jnp.hstack((x, condition))
-        transformer_params = self.autoregressive_mlp(nn_input)
+        transformer_params = self.masked_autoregressive_mlp(nn_input)
         transformer = self._flat_params_to_transformer(transformer_params)
         return transformer.transform_and_log_det(x)
 
@@ -116,7 +111,7 @@ class MaskedAutoregressive(AbstractBijection):
         """One 'step' in computing the inverse."""
         y, rank = init
         nn_input = y if condition is None else jnp.hstack((y, condition))
-        transformer_params = self.autoregressive_mlp(nn_input)
+        transformer_params = self.masked_autoregressive_mlp(nn_input)
         transformer = self._flat_params_to_transformer(transformer_params)
         x = transformer.inverse(y)
         x = y.at[rank].set(x[rank])
@@ -133,3 +128,43 @@ class MaskedAutoregressive(AbstractBijection):
         transformer_params = jnp.reshape(params, (dim, -1))
         transformer = eqx.filter_vmap(self.transformer_constructor)(transformer_params)
         return Vmap(transformer, in_axis=eqx.if_array(0))
+
+
+def masked_autoregressive_mlp(
+    in_ranks: ArrayLike,
+    hidden_ranks: ArrayLike,
+    out_ranks: ArrayLike,
+    **kwargs,
+) -> eqx.nn.MLP:
+    """Returns an equinox multilayer perceptron, with autoregressive masks.
+
+    The weight matrices are wrapped using :class:`~flowjax.wrappers.Where`, which
+    will apply the masking when :class:`~flowjax.wrappers.unwrap` is called on the MLP.
+    For details of how the masks are formed, see https://arxiv.org/pdf/1502.03509.pdf.
+
+    Args:
+        in_ranks: The ranks of the inputs.
+        hidden_ranks: The ranks of the hidden dimensions.
+        out_ranks: The ranks of the output dimensions.
+        **kwargs: Key word arguments passed to equinox.nn.MLP.
+    """
+    in_ranks, hidden_ranks, out_ranks = (
+        jnp.asarray(a, jnp.int32) for a in (in_ranks, hidden_ranks, out_ranks)
+    )
+    mlp = eqx.nn.MLP(
+        in_size=len(in_ranks),
+        out_size=len(out_ranks),
+        width_size=len(hidden_ranks),
+        **kwargs,
+    )
+    ranks = [in_ranks, *[hidden_ranks] * mlp.depth, out_ranks]
+
+    masked_layers = []
+    for i, linear in enumerate(mlp.layers):
+        mask = rank_based_mask(ranks[i], ranks[i + 1], eq=i != len(mlp.layers) - 1)
+        masked_linear = eqx.tree_at(
+            lambda linear: linear.weight, linear, Where(mask, linear.weight, 0)
+        )
+        masked_layers.append(masked_linear)
+
+    return eqx.tree_at(lambda mlp: mlp.layers, mlp, replace=tuple(masked_layers))
