--- flowjax/nn/__init__.py
@@ -1,9 +1,0 @@
-"""Contains invertible neural network architectures of use for constructing flows."""
-
-from .block_autoregressive import BlockAutoregressiveLinear
-from .masked_autoregressive import autoregressive_mlp
-
-__all__ = [
-    "autoregressive_mlp",
-    "BlockAutoregressiveLinear",
-]
