+++ tests/test_masks.py
@@ -7,18 +7,18 @@ def test_rank_based_mask():
     in_ranks = jnp.arange(2)
     out_ranks = jnp.array([0, 1, 1, 2])
 
-    expected_mask = jnp.array([[0, 0], [1, 0], [1, 0], [1, 1]], dtype=jnp.int32)
+    expected_mask = jnp.array(
+        [[False, False], [True, False], [True, False], [True, True]]
+    )
 
     mask = rank_based_mask(in_ranks, out_ranks)
-    assert mask.dtype == jnp.int32
     assert jnp.all(expected_mask == mask)
 
     in_ranks = jnp.array([0, 0, 1, 1])
     out_ranks = jnp.array([0, 1])
 
-    expected_mask = jnp.array([[0, 0, 0, 0], [1, 1, 0, 0]], dtype=jnp.int32)
+    expected_mask = jnp.array([[False] * 4, [True, True, False, False]])
     mask = rank_based_mask(in_ranks, out_ranks)
-    assert mask.dtype == jnp.int32
     assert jnp.all(expected_mask == mask)
 
 
@@ -26,7 +26,6 @@ def test_block_tril_mask():
     args = [(1, 2), 3]
     expected = jnp.array([[1, 1, 0, 0, 0, 0], [1, 1, 1, 1, 0, 0], [1, 1, 1, 1, 1, 1]])
     mask = block_tril_mask(*args)
-    assert mask.dtype == jnp.int32
     assert jnp.all(expected == mask)
 
     expected = jnp.array([[1, 1, 1, 1, 0, 0], [1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 1, 1]])
@@ -45,5 +44,4 @@ def test_block_diag_mask():
     args = [(1, 2), 3]
     expected = jnp.array([[1, 1, 0, 0, 0, 0], [0, 0, 1, 1, 0, 0], [0, 0, 0, 0, 1, 1]])
     mask = block_diag_mask(*args)
-    assert mask.dtype == jnp.int32
     assert jnp.all(expected == mask)
