+++ tests/test_wrappers.py
@@ -3,7 +3,7 @@ import jax.numpy as jnp
 import pytest
 
 from flowjax.bijections import Exp
-from flowjax.wrappers import BijectionReparam, Diagonal, unwrap
+from flowjax.wrappers import BijectionReparam, unwrap
 
 
 def test_BijectionReparam():
@@ -15,12 +15,3 @@ def test_BijectionReparam():
     wrapped = BijectionReparam(param, Exp())
     assert pytest.approx(unwrap(wrapped)) == param
     assert pytest.approx(wrapped.arr) == jnp.log(param)
-
-
-def test_Diagonal():
-
-    diag = Diagonal(jnp.ones((2, 2)))  # This should unwrap to 3D unlike jnp.diag
-    assert unwrap(diag).shape == (2, 2, 2)
-
-    expected = jnp.array([[[1.0, 0.0], [0.0, 1.0]], [[1.0, 0.0], [0.0, 1.0]]])
-    assert pytest.approx(unwrap(diag)) == expected
